Write 'apply_vect_func' function
Applies a function to every inner vector of a 2D vector. Type of vector
is generic but the function must take a vector and return a vector. For
this to work with the 'centre_vector' function it cannot be generic and
now take a double vector and returns a double vector.